requirements.txt
Django==1.11.4
django-ckeditor==5.3.0
django-js-asset==0.1.1
psycopg2==2.7.3.1
pytz==2017.2
